requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
fastapi
uvicorn
httpx
python-multipart
beautifulsoup4
ddgs>=4.0.0
pydantic>=2.0.0
lxml>=4.9.0
python-dotenv